+++ .gitignore
@@ -1,0 +1,4 @@
+
+.project
+.pydevproject
+*.pyc
